Add refactoring
