+++ .idea/workspace.xml
@@ -2,10 +2,12 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="c405ffb5-2ae5-41b2-92a1-1bd975c4e4fc" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/mocks/service/mockUserService.go" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/app/utilityFunctions.go" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/authMiddleware.go" beforeDir="false" afterPath="$PROJECT_DIR$/app/authMiddleware.go" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/app/countryHandlers.go" beforeDir="false" afterPath="$PROJECT_DIR$/app/countryHandlers.go" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/app/userHandlers.go" beforeDir="false" afterPath="$PROJECT_DIR$/app/userHandlers.go" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/service/userService.go" beforeDir="false" afterPath="$PROJECT_DIR$/service/userService.go" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/logger/logger.go" beforeDir="false" afterPath="$PROJECT_DIR$/logger/logger.go" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
