+++ logger/logger.go
@@ -16,9 +16,11 @@ func Info(message string, fields ...zap.Field) {
 	log.Info(message, fields...)
 }
 
+/*
 func Debug(message string, fields ...zap.Field) {
 	log.Info(message, fields...)
 }
+ */
 
 func Error(message string, fields ...zap.Field) {
 	log.Info(message, fields...)
